==> notebooks/1.0-mt-training-svm.ipynb <==
import glob
import csv
import pandas as pd
import re

csv_path = ""
data_path = "../data/interim/6_stages/"
files = glob.glob(f"{data_path}/*uncalibrated.csv")
data_uncalibrated = [pd.read_csv(file, delimiter=',', index_col=0) for file in files]
classes = [int(re.search('stage_.+?', filename).group(0).split('_')[-1]) for filename in files]
print(len(classes), classes)
print(data_uncalibrated[0])
print(len(data_uncalibrated))

from scipy.fft import fft, fftfreq, rfft, rfftfreq
import numpy as np
def fourier(dataframe, start=5, end=200):
    yf_xu = rfft(dataframe['X_UnCal'].to_numpy())
    yf_yu = rfft(dataframe['Y_UnCal'].to_numpy())
    yf_zu = rfft(dataframe['Z_UnCal'].to_numpy())
    x = dataframe['time_uncalibrated'].to_numpy()
    x = x - x[0]
    xf = rfftfreq(len(x), (x[1] - x[0]) / 1000)
    # this should be done parallel, not sequential
    return np.concatenate((yf_xu[start:end], yf_yu[start:end], yf_zu[start:end]))



X = [fourier(case, start=5, end=30) for case in data_uncalibrated]
# filter samples that are too small
index = [i  for i in range(len(X)) if len(X[i]) == 75]
X = [X[i] for i in index]
y = [classes[i] for i in index]

# to real values
X = np.real(X)
print(len(X))

from sklearn.model_selection import train_test_split
from sklearn import svm
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.33, random_state=42)
print(X_train.shape)
clf = svm.SVC(decision_function_shape='ovo')
clf.fit(X_train, y_train)
y_pred = clf.predict(X_test)
print(sum([1 for i in range(len(y_pred)) if y_pred[i] == y_test[i]])/len(y_pred))
print(y_pred)
print(y_test)
