==> notebooks/2.0-mt-json-to-csv.ipynb <==
import json
import pandas as pd
import glob
import uuid
import os
data_path = "../data/raw/v4"
interim_data_path = "../data/interim/v4"
if not os.path.isdir(interim_data_path):
    print("Created dir", interim_data_path)
    os.makedirs(interim_data_path)


files = glob.glob(f"{data_path}/Data*.json")
print(files)



for i, file_name in enumerate(files):
    # print(file_name)
    data = json.load(open(file_name, "r"))[0]

    # reading data
    first_timestamp_miliseconds = data["time_UnCal"][0]
    last_timestamp_miliseconds = data["time_UnCal"][-1]
    time_diff_miliseconds =  last_timestamp_miliseconds - first_timestamp_miliseconds
   
    df_uncalibrated = pd.DataFrame({k: data[k] for k in ["X_UnCal", "Y_UnCal", "Z_UnCal", "X_Bias", "Y_Bias", "Z_Bias", "time_UnCal", "Accuracy"]})

    
    df_uncalibrated_features = list(df_uncalibrated.columns)
    device_id = data['DEVICE_ID']
    stage = data["stage"]
    duration_seconds = time_diff_miliseconds / 1000
    sample_rate = df_uncalibrated.size / duration_seconds
    
    user_id = file_name.split('_')[-1].split('.')[0]
    df_uncalibrated.to_csv(f"{interim_data_path}/{device_id or 'no_id'}-stage_{stage}-{user_id}-uncalibrated.csv")
    print(f"{i}/{len(files)}")


# create metadata file
metadata = json.load(open(f"{data_path}/metadata.json","r"))
print(metadata)
metadata['format'] = "interim"

# Serializing json
json_object = json.dumps(metadata, indent=4)
 
# Writing to sample.json
with open(f"{interim_data_path}/metadata.json", "w") as outfile:
    outfile.write(json_object)
