==> notebooks/2.0-mt-training-svm_PCA_RF.ipynb <==
# Training on 2 classes

In this experiment, we trained on 2 classes, `all_off` (0), where all the communication is off, and `all_on` (1), where everything is on.

We tested in following conditions:

Number of test cases: 1

Location: inside, in kitchen, next to a computer and several apliances 

Communication modules (internal producers of electromagnetic emissions)
- wifi
- bluetooth
- mobile antenna

Unfortunately, due to miscommunication in our team, `all_off` phase did not completely turn off mobile antenna (we turned off mobile data, but did not turn on airplane mode)

import glob
import csv
import pandas as pd
import re
import numpy as np

csv_path = ""
data_path = "../data/interim/2_stages"
files = glob.glob(f"{data_path}/*uncalibrated.csv")
data_uncalibrated = [pd.read_csv(file, delimiter=',', index_col=0) for file in files]
classes = [int(re.search('stage_.+?', filename).group(0).split('_')[-1]) for filename in files]
n_of_train_cases = len(classes)
print(f"{n_of_train_cases = }", classes)
print(data_uncalibrated[0])
print(len(data_uncalibrated))

We are using uncalibrated sensor values. Since the sampling rate during `all_on` phase was 2 times higher, we corrected the data.

**Estimated sampling rate: 200Hz**

def to_numpy(df: pd.DataFrame):
    return df[['X_UnCal', 'Y_UnCal', 'Z_UnCal']].values

class_0_3d = to_numpy(data_uncalibrated[0])
class_1_3d = to_numpy(data_uncalibrated[1][::2])

# class_0_3d
print(len(class_0_3d))
print(len(class_1_3d))

We generated sequences of fixed window size (1000 readings, 5 seconds) with stride 100

def subsequences(ts, window, stride=2):
    assert ts.shape[1] == 3
    return np.lib.stride_tricks.sliding_window_view(ts, (window, ts.shape[1]))[:,0,:,:][::stride]

test = np.array([
    [0, 1, 2, 3, 4, 5, 6, 7, 8],
    [9, 10, 11, 12, 13, 14, 15, 16, 17], 
    [18,19,20,21,22,23,24,25,26]]).transpose()
print(test)
result = subsequences(test, window=4, stride=3)
print(result.shape)
print(result)

We have a dataset of 50 test cases from every testing mode (`all_on` and `all_off`), every dataset a matrix `(1000, 3)` as `(frequenceies, spatial_dimension_xyz)`

test_cases_0 = subsequences(class_0_3d, window=1000, stride=100)
test_cases_1 = subsequences(class_1_3d, window=1000, stride=100)
print(test_cases_0.shape)
print(test_cases_1.shape)

We are using `rfftn`, which is `N-dimensional discrete Fourier Transform for real input`

from scipy.fft import rfftn
def fourier(df, treshold=10):
    return rfftn(df, axes=0, norm="forward")[treshold:]

from scipy.fft import rfftn

rfftn(class_0_3d, axes=0, norm="forward")

from scipy.fft import rfftn
cut = 20

X_0 = np.array([fourier(test_case, cut) for test_case in test_cases_0])
X_1 = np.array([fourier(test_case, cut) for test_case in test_cases_1])

Y_0 = np.zeros(X_0.shape[0])
Y_1 = np.ones(X_1.shape[0])

print(Y_0.shape)
print(Y_1.shape)

print(X_0.shape)
print(X_1.shape)

Fourier transform returns complex values, we keep just real ones

X = np.concatenate((X_0, X_1), axis=0)
Y = np.concatenate((Y_0, Y_1), axis=0)

# to real values
X = np.real(X)

print(X.shape)
print(Y.shape)


2 methods for projecting into 1-dim test cases:
- just pick one of 3 dims
- flatten


# pick dimension
X_dim_0 = X[:,:,1]
print(X_dim_0.shape)

# flatten
X_flatten = np.array([el.flatten('F') for el in X])
print(X_flatten.shape)

X_to_train = X_dim_0
# X_to_train = X_flatten

## one dimension, fourier, svm

from sklearn.model_selection import train_test_split
from sklearn.metrics import confusion_matrix
from sklearn import svm

X_train, X_test, y_train, y_test = train_test_split(X_to_train, Y, test_size=0.33, random_state=42)
print(y_train.shape)
clf = svm.SVC(decision_function_shape='ovo')
clf.fit(X_train, y_train)
y_pred = clf.predict(X_test)
print(sum([1 for i in range(len(y_pred)) if y_pred[i] == y_test[i]])/len(y_pred))

classes_real = ["all_off", "all_on"]
classes_pred = ["all_off_predicted", "all_on_predicted"]
pd.DataFrame(confusion_matrix(y_test, y_pred), classes_real, classes_pred)

## 2 classes, fourier + PCA (1443 -> 100 dim)  + SVM

from sklearn.decomposition import PCA

print(X_flatten.shape)

pca = PCA(n_components=100)
X_PCA =  pca.fit(X_flatten).transform(X_flatten)
print(X_PCA.shape)

from sklearn.model_selection import train_test_split
from sklearn.metrics import confusion_matrix
from sklearn import svm

X_train, X_test, y_train, y_test = train_test_split(X_PCA, Y, test_size=0.33, random_state=42)
print(y_train.shape)
clf = svm.SVC(decision_function_shape='ovo')
clf.fit(X_train, y_train)
y_pred = clf.predict(X_test)
print(sum([1 for i in range(len(y_pred)) if y_pred[i] == y_test[i]])/len(y_pred))

classes_real = ["all_off", "all_on"]
classes_pred = ["all_off_predicted", "all_on_predicted"]
pd.DataFrame(confusion_matrix(y_test, y_pred), classes_real, classes_pred)

## 2 classes, Fourier + flattening + RF

Fourier analysis. Cut first 20 frequencies. Flatten matrix (freq, 3). Random forrest of depth 2

from sklearn.ensemble import RandomForestClassifier
X_train, X_test, y_train, y_test = train_test_split(X_flatten, Y, test_size=0.33)
clf = RandomForestClassifier(max_depth=2)
clf.fit(X_train, y_train)
y_pred = clf.predict(X_test)
print(sum([1 for i in range(len(y_pred)) if y_pred[i] == y_test[i]])/len(y_pred))

classes_real = ["all_off", "all_on"]
classes_pred = ["all_off_predicted", "all_on_predicted"]
pd.DataFrame(confusion_matrix(y_test, y_pred), classes_real, classes_pred)

## Remarks
- RF results could be overfitted (entire dataset from one 60min sampling session)
- dimensionality reduction on test-case level is TBD. Could do PCA on every single test case
- this should be tested on multiple samplings, multiple phones
