==> notebooks/3.0-mt-feature_selection_and_training-SVM_RF.ipynb <==
# Training on 2 classes (V4 dataset)

In this experiment, we trained on 2 classes, `all_off` (5), where all the communication is off, and `all_on` (6), where everything is on.

We tested in following conditions:

Number of test cases: 12 (6 from each class)

Locations
- park
- kitchen
- room

Communication modules (internal producers of electromagnetic emissions)
- wifi
- bluetooth
- mobile antenna

Traffic on

import os
import pandas as pd
import glob
import json
version = "v4"
dataset_path = f"../data/processed/"
filename = f"{version}-2_stages-full_statistical_with_window.csv"

import csv
import pandas as pd

df = pd.read_csv(f"{dataset_path}/{filename}")
df


df.columns

df['Name'].to_list()

# Feature selection
- corelation between feature and label
- 

correlation = df.iloc[:, 3:].apply(lambda x: df['Label'].corr(x, method="pearson")).abs().sort_values(ascending=False)
correlation

# select features with correlation greater then 0.1
selection = correlation[correlation > 0.1].index
df[selection]

- test: all subsequences from 'milenkoParkWithTraffic1 _'
- train: everything else

# one session test, all other sessions train

test_case = 'milenkosRoomNearDevicesWithTraffic_'
# test_case = 'milenkosRoomNearDevicesWithTraffic_'
test = df[df['Name'].str.contains(test_case)]
train = df[~df['Name'].str.contains(test_case)]
test

y_test = test['Label'].to_numpy()
X_test = test[selection].to_numpy()
y_train = train['Label'].to_numpy()
X_train = train[selection].to_numpy()


## SVM

from sklearn.model_selection import train_test_split
from sklearn.metrics import confusion_matrix
from sklearn import svm

print(y_train.shape)
clf = svm.SVC(decision_function_shape='ovo')
clf.fit(X_train, y_train)
y_pred = clf.predict(X_test)
print(sum([1 for i in range(len(y_pred)) if y_pred[i] == y_test[i]])/len(y_pred))

classes_real = ["all_off", "all_on"]
classes_pred = ["all_off_predicted", "all_on_predicted"]
pd.DataFrame(confusion_matrix(y_test, y_pred), classes_real, classes_pred)

## RF

from sklearn.ensemble import RandomForestClassifier
clf = RandomForestClassifier(max_depth=2)
clf.fit(X_train, y_train)
y_pred = clf.predict(X_test)
print(sum([1 for i in range(len(y_pred)) if y_pred[i] == y_test[i]])/len(y_pred))

classes_real = ["all_off", "all_on"]
classes_pred = ["all_off_predicted", "all_on_predicted"]
pd.DataFrame(confusion_matrix(y_test, y_pred), classes_real, classes_pred)

# sessions from the same location
- train: one session
- test: another session on the same location

test_case = 'milenkoParkWithTraffic_'
train_case = 'milenkoParkWithTraffic1 _'
test = df[df['Name'].str.contains(test_case)]
train = df[df['Name'].str.contains(train_case)]

y_test = test['Label'].to_numpy()
X_test = test[selection].to_numpy()
y_train = train['Label'].to_numpy()
X_train = train[selection].to_numpy()


## SVM

from sklearn.model_selection import train_test_split
from sklearn.metrics import confusion_matrix
from sklearn import svm

print(y_train.shape)
clf = svm.SVC(decision_function_shape='ovo')
clf.fit(X_train, y_train)
y_pred = clf.predict(X_test)
print(sum([1 for i in range(len(y_pred)) if y_pred[i] == y_test[i]])/len(y_pred))

classes_real = ["all_off", "all_on"]
classes_pred = ["all_off_predicted", "all_on_predicted"]
pd.DataFrame(confusion_matrix(y_test, y_pred), classes_real, classes_pred)

## RF

from sklearn.ensemble import RandomForestClassifier
clf = RandomForestClassifier(max_depth=2)
clf.fit(X_train, y_train)
y_pred = clf.predict(X_test)
print(sum([1 for i in range(len(y_pred)) if y_pred[i] == y_test[i]])/len(y_pred))

classes_real = ["all_off", "all_on"]
classes_pred = ["all_off_predicted", "all_on_predicted"]
pd.DataFrame(confusion_matrix(y_test, y_pred), classes_real, classes_pred)

# Another location
- Test: one session, train all others, but not on the same location

# test_case = 'milenkoPark'
test_case = 'milenkosKitchenNotNearDevices'

# 30% test
test = df[df['Name'].str.contains(test_case)]
train = df[~df['Name'].str.contains(test_case)]


y_test = test['Label'].to_numpy()
X_test = test[selection].to_numpy()
y_train = train['Label'].to_numpy()
X_train = train[selection].to_numpy()

## SVM

from sklearn.model_selection import train_test_split
from sklearn.metrics import confusion_matrix
from sklearn import svm

print(y_train.shape)
clf = svm.SVC(decision_function_shape='ovo')
clf.fit(X_train, y_train)
y_pred = clf.predict(X_test)
print(sum([1 for i in range(len(y_pred)) if y_pred[i] == y_test[i]])/len(y_pred))

classes_real = ["all_off", "all_on"]
classes_pred = ["all_off_predicted", "all_on_predicted"]
pd.DataFrame(confusion_matrix(y_test, y_pred), classes_real, classes_pred)

from sklearn.ensemble import RandomForestClassifier
clf = RandomForestClassifier(max_depth=7)
clf.fit(X_train, y_train)
y_pred = clf.predict(X_test)
print(sum([1 for i in range(len(y_pred)) if y_pred[i] == y_test[i]])/len(y_pred))

classes_real = ["all_off", "all_on"]
classes_pred = ["all_off_predicted", "all_on_predicted"]
pd.DataFrame(confusion_matrix(y_test, y_pred), classes_real, classes_pred)

## Remarks
- it does not work if not on already seen location
