==> notebooks/3.0-mt-generate-fourier-dataset.ipynb <==
import os
import pandas as pd
import glob
import json
data_path = "../data/interim/v3"
dataset_target_path = "../data/processed"
test_cases = [file.split("-")[2] for file in glob.glob(f"{data_path}/*stage_5*.csv")]
# List test cases in dataset
for i,test_case in enumerate(test_cases):
    print(i,test_case)

# read all data in dataset
data_off = {test_case: pd.read_csv(glob.glob(f"{data_path}/*stage_5*{test_case}*")[0]) for test_case in test_cases}
data_on = {test_case: pd.read_csv(glob.glob(f"{data_path}/*stage_6*{test_case}*")[0]) for test_case in test_cases}


def add_intensity(df: pd.DataFrame):
    """
    add Intensity as 2nd norm of 3D vector
    """
    df["Intensity"] = (df["X_UnCal"] ** 2 + df["Y_UnCal"] ** 2 + df["Z_UnCal"] ** 2) ** 0.5

for key in data_off.keys():
    add_intensity(data_off[key])
for key in data_on.keys():
    add_intensity(data_on[key])

from scipy.fft import fft, fftfreq
import numpy as np
from matplotlib import pyplot as plt

def spectral_image(df: pd.DataFrame, column = 'Intensity', duration = 60):
    N = len(df)
    sampling_rate = N // duration
    T = 1 / sampling_rate

    yf = fft(df[column].to_numpy())

    # just right side
    yff = 2.0/N * np.abs(yf[0:N//2])
    xf = fftfreq(N, T)[:N//2]

    bins = np.array(range(round(xf[-1])))
    inds = np.digitize(xf, bins)
    df = pd.DataFrame({"bins": inds, "fourier": yff})
    groupped = df.groupby("bins").sum()
    return groupped

def specter_to_data_row(name, label, df, vec_len, column = 'Intensity', duration = 60,): 
    specter_intensity = spectral_image(df, column, duration)['fourier'].to_numpy()[:vec_len]
    padded = np.zeros((vec_len))
    padded[:specter_intensity.shape[0]] = specter_intensity
    data_row = [name, label] + list(padded)
    return data_row

vec_len = 50
duration=60
columns = ["Name", "Label"] + list(range(1,vec_len+1))

data_on = pd.DataFrame([specter_to_data_row(name, 6, df, vec_len=50, column='Intensity') for name, df in data_on.items()], columns=columns)
data_off = pd.DataFrame([specter_to_data_row(name, 5, df, vec_len=50, column='Intensity') for name, df in data_off.items()], columns=columns)

data_fourier = data_on.append(data_off, ignore_index=True)
data_fourier = data_fourier.sort_values(by=['Label', 'Name'], ignore_index = True)
data_fourier

# save dataset
if not os.path.isdir(dataset_target_path):
    os.makedirs(dataset_target_path)

import time
# metadata
metadata = {
    "title": "Fourier dataset, 50 frequencies",
    "version": "v3",
    "description": "Dataset from v3 data, single device, without traffic, statistical features",
    "author": "Mihael",
    "places": [
        "room",
        "kitchen"
    ],
    "stages": 2,
    "traffic": False,
    "format": "processed",
    "created": int(time.time())
}
file_name = f"{metadata['version']}-{metadata['stages']}_stages-fourier"
data_fourier.to_csv(f"{dataset_target_path}/{file_name}.csv",index=True)
# Writing metadata
with open(f"{dataset_target_path}/{file_name}.json", "w") as outfile:
    outfile.write(json.dumps(metadata, indent=4))
