==> notebooks/3_0-mt-generate-full-fourier-statistical-dataset.ipynb <==
import os
import pandas as pd
import glob
import json
version = "v4"
data_path = f"../data/interim/{version}"
dataset_target_path = "../data/processed"


from lib import read_dataset
data_on_raw, data_off_raw = read_dataset(data_path, classes = 2)

data_on_raw[list(data_on_raw.keys())[0]]

from lib import add_intensity_to_dataset
add_intensity_to_dataset(data_on_raw)
add_intensity_to_dataset(data_off_raw)

for k,v in data_on_raw.items():
    print(v.size)
    break

from collections import ChainMap
def subsamples(df: pd.DataFrame, name: str, window=1000, stride=500):
    return {f"{name}_{i:03d}": df[start:start+window] for i, start in enumerate(range(0,df.shape[0]-window, stride))}

data_on = dict(ChainMap(*[subsamples(df, name) for name,df in data_on_raw.items()]))
data_off = dict(ChainMap(*[subsamples(df, name) for name,df in data_off_raw.items()]))




data_on[list(data_on.keys())[0]]

from lib import spectral_dataset, join_and_order_dataset
dataset_on = spectral_dataset(data_on, 6, vec_len=50, column='Intensity', experiment_duration=10)
dataset_off = spectral_dataset(data_off, 5, vec_len=50, column='Intensity', experiment_duration=10)
data_fourier = join_and_order_dataset([dataset_on, dataset_off])
data_fourier

statistics = data_fourier.iloc[:, 2:].apply(pd.DataFrame.describe, axis=1).iloc[:, 1:]
fourier_statistical = pd.concat([data_fourier[['Name', 'Label']], statistics], axis=1)
fourier_statistical.rename(lambda x: f"fourier_{x}", axis='columns', inplace=True)
fourier_statistical


from lib import statistical_dataset, join_and_order_dataset
# all_on is class 6
dataset_on = statistical_dataset(data_on, label=6)
# all_of is class 5
dataset_off = statistical_dataset(data_off, label=5)
data_statistical = join_and_order_dataset([dataset_on, dataset_off])
data_statistical

# join both datasets

dataset = pd.concat([data_statistical,fourier_statistical.iloc[:, 2:]], axis=1)
dataset

import time
from lib import save_dataset
# metadata
with open(f"{data_path}/metadata.json", 'r') as jsonfile:
    metadata = json.load(jsonfile)
    metadata['title'] = f"{version} full statistical dataset"
    metadata['description'] = 'Dataset with full statistical features and windowed test cases from '+metadata['description']
    metadata['format'] = "processed"
    metadata['author'] = 'Mihael'
    metadata['version'] = version
    metadata['created'] = int(time.time())
    save_dataset(dataset, dataset_target_path, dataset_name = 'full_statistical_with_window', metadata=metadata)
