==> notebooks/4.0-names.ipynb <==
import glob

import glob
data_path = "../data/raw/"
versions = ["v4", "v5"]
l = [glob.glob(f"{data_path}{version}/Data*") for version in versions]
file_names = [item for sublist in l for item in sublist]
test_names = [filename.split("_")[3][:-5] for filename in file_names]
print(len(test_names))
for name in test_names:
    print(name)
#print(test_names)



def remove_from_string(test_string):
    # remove comments
    remove_list = ["resorant", "Trafic", "Traffic", "With", "Truffic", "Eith", "DifferentOrentationPousesInRuns", "Orient"]
    for sub in remove_list:
        test_string = test_string.replace(sub, '')
    # remove numbers
    test_string = "".join(list(filter(lambda x: x.isalpha(), test_string)))
    return test_string
cleaned = [remove_from_string(name) for name in test_names]
for name, name_cleaned in sorted(zip(test_names, cleaned)):
    print(name_cleaned, name)

remove_from_string("LATERNAresorantWithTrafic")

for name in (sorted(list(set(cleaned)))):
    print(name)

place_mapping = {
    "NadasKitchen": "Kitchen1",
"MihasLivingRoom": "LivingRoom1",
"milenkosRoomBed": "DormRoom1",
"PiXmilenkosRoomTable": "DormRoom2",
"spelaMilenkosRoomTable": "DormRoom3",
"milenkoPark": "Park",
"LATERNA": "Restaurant",
"NadasLivingRoom": "LivingRoom2",
"milenkosRoomNearDevices": "DormRoom4",
"milenkosRoom": "DormRoom5",
"milenkosRoomTable": "DormRoom6",
"NadasBathroom": "Bathroom",
"spelaMilenkosKitchenTable": "Kitchen2",
"milenkosKitchenNotNearDevices": "Kitchen3"
}

for v in sorted(place_mapping.values()):
    print(f'"{v}",')
