==> notebooks/4_0-fourier_dataset_to_images.ipynb <==
import os
import pandas as pd
import numpy as np
import glob
import json
input_dataset_path = '../data/processed/v5-2_stages-fourier_window_big.csv'
out_path_off = '../data/processed/spectral/05'
out_path_on = '../data/processed/spectral/06'
if not os.path.exists(out_path_off):
    os.makedirs(out_path_off)
if not os.path.exists(out_path_on):
    os.makedirs(out_path_on)


df = pd.read_csv(input_dataset_path)
test_cases = list(set([name.split('_')[0] + '_' for name in df['name'].to_list()]))
test_cases

df_on = df[df['label'] == 6]
df_off = df[df['label'] == 5]

img = df_on[df_on['name'].str.contains(test_cases[0])].iloc[:,6:].to_numpy().transpose()
np.max(img)

a = []
for test_case in test_cases:
    df_temp = df_on[df_on['name'].str.contains(test_case)]
    device_id = df_temp['device_id'].iloc[0]
    a.append((device_id, df_temp.shape))
    #print(.shape)
sorted(a)

import matplotlib.pyplot as plt
import matplotlib.image

for test_case in test_cases:
    device_id = df_on[df_on['name'].str.contains(test_case)]['device_id'].iloc[0]
    location = df_on[df_on['name'].str.contains(test_case)]['location'].iloc[0]

    img = df_on[df_on['name'].str.contains(test_case)].iloc[:,6:].to_numpy().transpose()
    matplotlib.image.imsave(f"{out_path_on}/{test_case}{location}_{device_id}_on.png", img)
    
    img = df_off[df_off['name'].str.contains(test_case)].iloc[:,6:].to_numpy().transpose()
    matplotlib.image.imsave(f"{out_path_off}/{test_case}{location}_{device_id}_off.png", img)
