==> notebooks/4_0-generate_fourier_dataset.ipynb <==
import os
import pandas as pd
import glob
import json
versions = ["v4", "v5", "v6"]
dataset_target_path = "../data/processed"


from lib import read_dataset
from collections import ChainMap
data_on_raw_list = list()
data_off_raw_list = list()
for version in versions:
    data_path = f"../data/interim/{version}"
    # print(data_path)
    data_on_raw_tmp, data_off_raw_tmp = read_dataset(data_path, classes = 2)
    # print(len(data_off_raw_tmp.keys()))
    data_off_raw_list.append(data_off_raw_tmp)
    data_on_raw_list.append(data_on_raw_tmp)

data_on_raw = dict(ChainMap(*data_on_raw_list))
data_off_raw = dict(ChainMap(*data_off_raw_list))

for i,test_case in enumerate(data_off_raw):
             print(i,test_case)

from lib import normalize_dataset, standardize_dataset
columns = ['X_UnCal', 'Z_UnCal', 'Y_UnCal']#, 'Intensity']
#standardize_dataset(data_on_raw, columns)
#standardize_dataset(data_off_raw, columns)

#normalize_dataset(data_on_raw, columns)
#normalize_dataset(data_off_raw, columns)


from lib import add_intensity_to_dataset
add_intensity_to_dataset(data_on_raw)
add_intensity_to_dataset(data_off_raw)

list(data_on_raw.values())[0].describe()

from collections import ChainMap
def subsamples(df: pd.DataFrame, name: str, window=2000, stride=10):
    return {f"{name}_{i:03d}": df[start:start+window] for i, start in enumerate(range(0,df.shape[0]-window, stride))}

data_on = dict(ChainMap(*[subsamples(df, name) for name,df in data_on_raw.items()]))
data_off = dict(ChainMap(*[subsamples(df, name) for name,df in data_off_raw.items()]))

len(data_on.keys())

data_on[list(data_on.keys())[0]]

from lib import spectral_dataset
dataset_on_intensity = spectral_dataset(data_on, 6, vec_len=50, column='Intensity', experiment_duration=10)
dataset_on_X = spectral_dataset(data_on, 6, vec_len=50, column='X_UnCal', experiment_duration=10)
dataset_on_Y = spectral_dataset(data_on, 6, vec_len=50, column='Y_UnCal', experiment_duration=10)
dataset_on_Z = spectral_dataset(data_on, 6, vec_len=50, column='Z_UnCal', experiment_duration=10)
dataset_off_intensity = spectral_dataset(data_off, 5, vec_len=50, column='Intensity', experiment_duration=10)
dataset_off_X = spectral_dataset(data_off, 5, vec_len=50, column='X_UnCal', experiment_duration=10)
dataset_off_Y = spectral_dataset(data_off, 5, vec_len=50, column='Y_UnCal', experiment_duration=10)
dataset_off_Z = spectral_dataset(data_off, 5, vec_len=50, column='Z_UnCal', experiment_duration=10)

dataset_on_intensity.rename(lambda x: f"intensity_{x}", axis='columns', inplace=True)
dataset_on_X.rename(lambda x: f"X_{x}", axis='columns', inplace=True)
dataset_on_Y.rename(lambda x: f"Y_{x}", axis='columns', inplace=True)
dataset_on_Z.rename(lambda x: f"Z_{x}", axis='columns', inplace=True)

dataset_off_intensity.rename(lambda x: f"intensity_{x}", axis='columns', inplace=True)
dataset_off_X.rename(lambda x: f"X_{x}", axis='columns', inplace=True)
dataset_off_Y.rename(lambda x: f"Y_{x}", axis='columns', inplace=True)
dataset_off_Z.rename(lambda x: f"Z_{x}", axis='columns', inplace=True)

dataset_on = pd.concat([dataset_on_intensity, dataset_on_X.iloc[:,4:], dataset_on_Y.iloc[:,4:], dataset_on_Z.iloc[:,4:]], axis=1)
dataset_on = dataset_on.rename(columns={"intensity_name": "name", "intensity_location": "location", "intensity_device_id": "device_id", "intensity_label": "label"})

dataset_off = pd.concat([dataset_off_intensity, dataset_off_X.iloc[:,4:], dataset_off_Y.iloc[:,4:], dataset_off_Z.iloc[:,4:]], axis=1)
dataset_off = dataset_off.rename(columns={"intensity_name": "name", "intensity_location": "location", "intensity_device_id": "device_id", "intensity_label": "label"})

dataset_off

from lib import join_and_order_dataset
data_fourier = join_and_order_dataset([dataset_on, dataset_off])
data_fourier

import time
from lib import save_dataset
with open(f"{data_path}/metadata.json", 'r') as jsonfile:
    metadata = json.load(jsonfile)
    metadata['title'] = f"{version} fourier dataset"
    metadata['description'] = 'Fourier dataset with windowed test cases from '+metadata['description']
    metadata['format'] = "processed"
    metadata['author'] = 'Mihael'
    metadata['version'] = versions[-1]
    metadata['created'] = int(time.time())
save_dataset(data_fourier, dataset_target_path, dataset_name = 'fourier_int_XYZ', metadata=metadata)
