==> notebooks/5.0-mt-locations-per-dataset.ipynb <==
import json
import pandas as pd
import glob
import uuid
import os
import itertools
from lib import remove_from_string, location_mapping, device_mapping, rotations_mapping

datasets = ['v4', 'v5', 'v6']

results = []

data_path = f""
files = list(itertools.chain(*[glob.glob(f"../data/raw/{dataset}/Data*.json") for dataset in datasets]))
for file in files:
    filename = os.path.basename(file)[:-5].split('_')
    test_name = filename[-1]
    device_id = filename[1]
    stage = filename[2]
    location = location_mapping.get(remove_from_string(test_name))
    results.append([test_name, location, device_mapping.get(device_id), stage, rotations_mapping[location]])#, file])

test_cases = pd.DataFrame(results, columns=['test_name', 'location', 'device', 'label', 'rotations'])#, 'path'])

pd.set_option('display.max_rows', 100)
# list of test cases
test_cases.sort_values(["device", "location", "test_name", "label"], ignore_index=True, inplace=True)
test_cases

# number of samples per location and per device
dev_per_loc = test_cases[["device", "location", "rotations"]].groupby('device').value_counts()
print(dev_per_loc.reset_index().to_latex(index=False))

# number of samples per device
test_cases[["device"]].groupby('device').value_counts()

# number of test cases per class
test_cases[["label"]].groupby("label").value_counts()
