==> notebooks/5.0-mt-normalization.ipynb <==
# Experimenting with new normalization

- this playbook contains the code for the experimentation with new device

import os
import csv
import pandas as pd

version = "v6"
dataset_path = f"../data/processed/"

# files
filename_s = f"{version}-2_stages-full_statistical.csv"
filename_s_n = f"{version}-2_stages-full_statistical_normalized.csv"
filename_s_s = f"{version}-2_stages-full_statistical_standardized.csv"
filename_f = f"{version}-2_stages-fourier_int_XYZ.csv"
filename_f_n = f"{version}-2_stages-fourier_normalized_int_XYZ.csv"
filename_f_s = f"{version}-2_stages-fourier_standardized_XYZ.csv"


# datasets
df_statistical = pd.read_csv(f"{dataset_path}/{filename_s}")
df_statistical_normalized = pd.read_csv(f"{dataset_path}/{filename_s_n}")
df_statistical_standardized = pd.read_csv(f"{dataset_path}/{filename_s_s}")
df_spectral = pd.read_csv(f"{dataset_path}/{filename_f}")
df_spectral_normalized = pd.read_csv(f"{dataset_path}/{filename_f_n}")
df_spectral_standardized = pd.read_csv(f"{dataset_path}/{filename_f_s}")

from lib import device_mapping, reverse_device_mapping

# ------------------------------------------------
# choice of dataset

# df = df_statistical # 56, 67
# df = df_statistical_normalized  # 54 63
# df = df_statistical_standardized    # 66, 72
# df = df_spectral    # 74, 77
# df = df_spectral_normalized #77 78
# df = df_spectral_standardized   #89, 80
datasets = [df_spectral, df_spectral_normalized, df_spectral_standardized]
dataset_names = ["raw", "normalized", "standardized"]

# ------------------------------------------------
# choice of phone

# GooglePixel6:        03575768cc23b2df
# RedmiNote8PRO:       4aaf95a621ccf092
# SamsungGalaxyA51:    029a77f196804217
# SamsungGalaxyS6:     e08d976ac75c011e


# df = df[df['device_id'].isin(reverse_device_mapping(['RedmiNote8PRO']))]
# df = df[df['device_id'].isin(reverse_device_mapping(['SamsungGalaxyS6']))]
# df = df[df['device_id'].isin(reverse_device_mapping(['GooglePixel6', 'RedmiNote8PRO']))]
# df = df[df['device_id'].isin(reverse_device_mapping(['SamsungGalaxyA51', 'SamsungGalaxyS6']))]

# ------------------------------------------------
# choice of locations to include TOOD

# number of cases per device
# df['device_id'].value_counts()


def map_classes(x):
    if x == 5:
        return 1
    return 0
def map_back_classes(x):
    if x == 1:
        return 5
    return 6

df = datasets[0]
# 03575768cc23b2df
device = "GooglePixel6"
device_id = reverse_device_mapping(device)[0]
print(device_id)
df['device_id'] == device_id

from sklearn import svm
from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import f1_score
from lib import device_mapping

results_svm_all = []
results_rf_all = []
for device in ["GooglePixel6", "RedmiNote8PRO", "SamsungGalaxyA51","SamsungGalaxyS6"]:
    results_svm  = [device]
    results_RF  = [device]
    for i, dataset in enumerate(datasets):
        df = dataset
        
        dataset_name = dataset_names[i]
        # test is one device, train the rest
        test = df[df['device_id'] == reverse_device_mapping(device)[0]]
        train = df[df['device_id'] != reverse_device_mapping(device)[0]]

        # correlation and feature selection
        correlation = df.iloc[:, 6:].apply(lambda x: df['label'].corr(x, method="pearson")).abs().sort_values(ascending=False)
        selection = correlation[:10].index

        y_test = test['label'].to_numpy()
        X_test = test[selection].to_numpy()
        y_train = train['label'].to_numpy()
        X_train = train[selection].to_numpy()

        # SVM
        clf = svm.SVC(decision_function_shape='ovo')
        clf.fit(X_train, y_train)
        y_pred = clf.predict(X_test)
        acc_svm = sum([1 for i in range(len(y_pred)) if y_pred[i] == y_test[i]])/len(y_pred)
        f1_svm = f1_score(list(map(map_classes, y_test)), list(map(map_classes, y_pred)))

        # RF
        clf = RandomForestClassifier(max_depth=3)
        clf.fit(X_train, y_train)
        y_pred = clf.predict(X_test)
        acc_rf = sum([1 for i in range(len(y_pred)) if y_pred[i] == y_test[i]])/len(y_pred)
        f1_rf = f1_score(list(map(map_classes, y_test)), list(map(map_classes, y_pred)))

        device_name = device_mapping[test['device_id'].iloc[0]]
        location = test['location'].iloc[0]
        #results_temp = [device, acc_svm, f1_svm, acc_rf, f1_rf]
        results_svm.append(acc_svm)
        results_RF.append(acc_rf)
    print(results_svm)
    print(results_RF)
    results_svm_all.append(results_svm)
    results_rf_all.append(results_RF)

results_all_svm = pd.DataFrame(results_svm_all, columns=['Device', "raw", "normalized", "standardized"])
results_all_rf = pd.DataFrame(results_rf_all, columns=['Device', "raw", "normalized", "standardized"])
results_all_svm

results_all_svm

results_all_rf

results_all_svm.mean()

results_all_rf.mean()

print(results_all_svm.sort_values(['Device']).round(3).to_latex(index=False))

print(results_all_svm.sort_values(['Device']).round(3).to_latex(index=False))
